Move compass-query-bar to shared config
