+++ .depalignrc.json
@@ -123,10 +123,10 @@
       "^3.1.0",
       "^2.13.1"
     ],
-    "numeral": ["^2.0.6", "^1.5.6", "^1.5.3"],
+    "numeral": ["^2.0.6", "^1.5.6"],
     "mini-css-extract-plugin": ["^0.8.0", "^0.9.0"],
     "mongodb-query-util": ["^0.2.1", "^0.0.3"],
-    "ms": ["^2.1.2", "^0.7.1"],
+    "ms": ["^2.1.2"],
     "pluralize": ["^8.0.0", "^1.2.1"],
     "d3-timer": ["^1.0.3", "1.0.10", "1.0.7"],
     "react-click-outside": ["^3.0.1", "^2.2.0"],
