+++ packages/compass-query-bar/.eslintignore
@@ -1,0 +1,2 @@
+.nyc-output
+dist
