--- packages/compass-query-bar/.eslintrc
@@ -1,3 +1,0 @@
-{
-  "extends": ["mongodb-js/compass-plugin"]
-}
