--- packages/compass-query-bar/.gitignore
@@ -1,13 +1,0 @@
-.DS_Store
-npm-debug.log
-node_modules/
-lib/
-lib-cov
-coverage
-.lock-wscript
-.idea/
-*.iml
-.nvmrc
-.nyc_output
-*.swp
-src/assets/fonts/akzid*
