+++ packages/compass-query-bar/.mocharc.js
@@ -1,0 +1,1 @@
+module.exports = require('@mongodb-js/mocha-config-compass/compass-plugin');
