--- packages/compass-query-bar/.npmignore
@@ -1,20 +1,0 @@
-.babelrc
-.eslintrc
-.editorconfig
-.nycrc
-.prettierrc
-config/
-electron/
-karma.conf.js
-mocha-webpack.opts
-scripts/
-src/
-test/
-.nyc_output/
-coverage/
-.github/
-lib/index.html
-lib/index.js.map
-
-scripts/**/*
-!scripts/download-akzidenz.js
