--- packages/compass-query-bar/.prettierrc
@@ -1,3 +1,0 @@
-{
-  "singleQuote": true
-}
