--- packages/compass-query-bar/mocha-webpack.opts
@@ -1,6 +1,0 @@
---webpack-config config/webpack.test.config.js
---require jsdom-global/register
---require test/setup.js
---recursive
---ui bdd
---timeout 15000
