+++ packages/compass-query-bar/package.json
@@ -2,13 +2,7 @@
   "name": "@mongodb-js/compass-query-bar",
   "productName": "Query Bar plugin",
   "version": "7.22.0",
-  "apiVersion": "3.0.0",
   "description": "Renders a component for executing MongoDB queries through a GUI.",
-  "main": "lib/index.js",
-  "compass:main": "src/index.js",
-  "compass:exports": {
-    ".": "./src/index.js"
-  },
   "homepage": "https://github.com/mongodb-js/compass",
   "license": "SSPL",
   "author": {
@@ -19,29 +13,50 @@
     "url": "https://jira.mongodb.org/projects/COMPASS/issues",
     "email": "compass@mongodb.com"
   },
+  "publishConfig": {
+    "access": "public"
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/mongodb-js/compass.git"
   },
+  "files": [
+    "dist"
+  ],
+  "main": "dist/index.js",
+  "compass:main": "src/index.js",
+  "exports": {
+    "require": "./dist/index.js",
+    "browser": "./dist/browser.js"
+  },
+  "compass:exports": {
+    ".": "./src/index.js"
+  },
   "scripts": {
-    "clean": "rimraf lib",
-    "precompile": "npm run clean",
-    "compile": "cross-env NODE_ENV=production webpack --config ./config/webpack.prod.config.js",
-    "start": "webpack-dev-server --config ./config/webpack.dev.config.js",
-    "start:watch": "npm run clean && webpack --config ./config/webpack.watch.config.js",
-    "test": "cross-env NODE_ENV=test mocha-webpack \"./src/**/*.spec.js\" \"./test/**/*.js\"",
-    "test-watch": "cross-env NODE_ENV=test mocha-webpack \"./src/**/*.spec.js\" \"./test/**/*.js\" --watch",
-    "test-electron": "xvfb-maybe cross-env NODE_ENV=test karma start",
-    "test-cov": "nyc npm run test",
-    "check": "npm run lint && npm run depcheck",
     "prepublishOnly": "npm run compile && compass-scripts check-exports-exist",
-    "lint": "eslint \"./src/**/*.{js,jsx}\" \"./test/**/*.js\" \"./electron/**/*.js\" \"./config/**/*.{js,jsx}\"",
+    "compile": "npm run webpack -- --mode production",
+    "prewebpack": "rimraf ./dist",
+    "webpack": "webpack-compass",
+    "start": "npm run webpack serve -- --mode development",
+    "analyze": "npm run webpack -- --mode production --analyze",
+    "typecheck": "tsc --noEmit",
+    "eslint": "eslint",
+    "prettier": "prettier",
+    "lint": "npm run eslint .",
+    "lint-skipped-for-now": "&& npm run prettier -- --check .",
     "depcheck": "depcheck",
-    "test-ci": "npm run test",
-    "test-ci-electron": "echo \"TODO(COMPASS-5555): These tests are broken and disabled for now\"",
-    "bootstrap": "npm run compile"
+    "check": "npm run typecheck && npm run lint && npm run depcheck",
+    "check-ci": "npm run check",
+    "test": "mocha",
+    "test-electron": "xvfb-maybe electron-mocha --no-sandbox",
+    "test-cov": "nyc -x \"**/*.spec.*\" --reporter=lcov --reporter=text --reporter=html npm run test",
+    "test-watch": "npm run test -- --watch",
+    "test-ci": "npm run test-cov",
+    "test-ci-electron": "npm run test-electron",
+    "reformat": "npm run prettier -- --write ."
   },
   "peerDependencies": {
+    "@mongodb-js/compass-components": "^0.15.0",
     "debug": "*",
     "hadron-react-components": "^5.15.0",
     "mongodb-query-parser": "^2.4.6",
@@ -54,63 +69,33 @@
     "reflux-state-mixin": "github:mongodb-js/reflux-state-mixin"
   },
   "devDependencies": {
-    "@babel/cli": "^7.14.3",
-    "@babel/core": "^7.14.3",
-    "@babel/plugin-proposal-decorators": "^7.14.2",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/preset-env": "^7.14.2",
-    "@babel/preset-react": "^7.13.13",
-    "@babel/register": "^7.13.16",
     "@mongodb-js/compass-components": "^0.15.0",
-    "autoprefixer": "^9.7.4",
-    "babel-loader": "^8.2.2",
-    "chai": "^4.1.2",
+    "@mongodb-js/eslint-config-compass": "^0.9.0",
+    "@mongodb-js/mocha-config-compass": "^0.10.0",
+    "@mongodb-js/prettier-config-compass": "^0.5.0",
+    "@mongodb-js/tsconfig-compass": "^0.6.0",
+    "@mongodb-js/webpack-config-compass": "^0.9.0",
+    "chai": "^4.2.0",
     "chai-enzyme": "^1.0.0-beta.0",
     "cheerio": "^1.0.0-rc.2",
     "classnames": "^2.2.6",
-    "core-js": "^3.12.1",
     "cross-env": "^7.0.0",
-    "css-loader": "^4.3.0",
     "debug": "4.3.0",
     "depcheck": "^1.4.1",
     "electron": "^13.5.1",
     "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
     "eslint": "^7.25.0",
-    "eslint-config-mongodb-js": "^5.0.3",
-    "extract-text-webpack-plugin": "^4.0.0-beta.0",
-    "file-loader": "^5.1.0",
     "font-awesome": "^4.7.0",
     "hadron-app-registry": "^8.12.0",
     "hadron-react-components": "^5.15.0",
-    "html-webpack-plugin": "^3.2.0",
-    "htmlhint": "^0.11.0",
-    "ignore-loader": "^0.1.2",
-    "inject-loader": "^4.0.1",
-    "istanbul-instrumenter-loader": "^3.0.1",
-    "jsdom": "^16.7.0",
-    "jsdom-global": "^3.0.2",
-    "karma": "^6.3.4",
-    "karma-chai": "^0.1.0",
-    "karma-chai-sinon": "^0.1.5",
-    "karma-electron": "^7.0.0",
-    "karma-mocha": "^1.3.0",
-    "karma-mocha-reporter": "^2.2.5",
-    "karma-sinon": "^1.0.5",
-    "karma-sourcemap-loader": "^0.3.8",
-    "karma-webpack": "^4.0.2",
-    "less": "^3.11.1",
-    "less-loader": "^7.3.0",
     "mocha": "^7.0.1",
-    "mocha-webpack": "^2.0.0-beta.0",
     "mongodb-ace-autocompleter": "^0.13.0",
-    "mongodb-connection-model": "^21.17.0",
     "mongodb-query-parser": "^2.4.6",
     "mongodb-reflux-store": "^0.0.1",
     "mongodb-schema": "^8.2.5",
     "ms": "^2.1.2",
     "numeral": "^2.0.6",
-    "nyc": "^15.0.0",
+    "nyc": "^15.1.0",
     "object-diff": "0.0.4",
     "peer-deps-externals-webpack-plugin": "^1.0.4",
     "pluralize": "^8.0.0",
@@ -120,23 +105,13 @@
     "react-bootstrap": "^0.32.1",
     "react-dom": "^16.14.0",
     "react-fontawesome": "^1.6.1",
-    "react-hot-loader": "^4.13.0",
     "reflux": "^0.4.1",
     "reflux-state-mixin": "github:mongodb-js/reflux-state-mixin",
     "resolve": "^1.15.1",
     "rimraf": "^3.0.1",
     "semver": "^5.4.1",
-    "shebang-loader": "^0.0.1",
-    "sinon": "^8.1.1",
+    "sinon": "^9.2.3",
     "sinon-chai": "^3.4.0",
-    "style-loader": "^2.0.0",
-    "url-loader": "^3.0.0",
-    "webpack": "^4.46.0",
-    "webpack-bundle-analyzer": "^3.6.0",
-    "webpack-cli": "^3.3.12",
-    "webpack-dev-server": "^3.11.2",
-    "webpack-merge": "^4.2.2",
-    "webpack-node-externals": "^3.0.0",
     "xvfb-maybe": "^0.2.1"
   },
   "dependencies": {
